Fix Interop TCs to use long long instead of long (dotnet/coreclr#5054)
Some TCs were using long for 64bit variable.
While it is correct on x86, long is 32bit on ARM.
Fix TCs to use long long instead so that they can pass on ARM as well.

Fix dotnet/coreclr#5053


Commit migrated from https://github.com/dotnet/coreclr/commit/b7704374e683a8d3644a0640010936a22296967a

+++ src/coreclr/tests/src/Interop/common/types.h
@@ -33,10 +33,10 @@ typedef WCHAR OLECHAR;
 
 typedef unsigned int UINT_PTR;
 
-typedef unsigned long ULONG64;
+typedef unsigned long long ULONG64;
 typedef double DOUBLE;
 typedef float FLOAT;
-typedef signed long LONG64, *PLONG64;
+typedef signed long long LONG64, *PLONG64;
 typedef int INT, *LPINT;
 typedef unsigned int UINT;
 typedef char CHAR, *PCHAR;
